receitas/python/verivox/requirements.txt
Appium-Python-Client>=1.0.2
pytest>=6.2.2
pytest-html-reporter>=0.2.3